Fix up possible memory leaks in libdeltacloud.
As pointed out by coverity, the parse_instance_xml() function
could leak memory if there was more than one XML element named
name, owner_id, image_href, relam_href, state, hardware_profile,
actions, public_address, or private_addresses.  Fix this by
using XPath expressions instead of a loop iterating over the
nodes.

Note that this does *not* give us the flexibility of supporting multiple
elements of the same name.  The old code would leak memory and only
return the last of the elements if this happened; the new code will
only return the first of the elements, which I consider an improvement.

The situation is unclear with respect to multiple elements with the
same name.  We should probably raise an error for invalid XML,
but for now we don't.

Signed-off-by: Chris Lalancette <clalance@redhat.com>