Don't overwrite curl errors.
When detailed curl error reporting was introduced, removal
of the old "get/post_url_error" generic reporting was not
removed.  This means that if curl failed to do something,
it would log the error, which would then immediately get
overwritten by the more generic error.  Remove the generic
errors and only rely on the curl errors for information.

Signed-off-by: Chris Lalancette <clalance@redhat.com>