Allow empty data return from curl actions.
The deltacloud API has been changed so that the return value
from POST can be (and most likely will be) empty.  However,
they still need to be able to succeed.  Therefore, change
the do_get_post_url() function to return a success code,
but also have a parameter to get the return value from
GETs.

Signed-off-by: Chris Lalancette <clalance@redhat.com>