Pass a parameter list to deltacloud_create_instance.
Up until now, every time we have needed to add a new
parameter to pass to deltacloud_create_instance(), we've
had to break the API to do that.  Not only is this annoying,
it is also error-prone.

To fix that, change deltacloud_create_instance to expect
an array of struct deltacloud_create_parameter.  Each of
these parameters has a name and a value, and
deltacloud_create_instance will properly build these into
the POST to create_instance.

We also introduce deltacloud_prepare_parameter,
deltacloud_allocate_parameter, deltacloud_free_parameter_value,
and deltacloud_free_parameter helper functions to allocate
parameters for the user.

Signed-off-by: Chris Lalancette <clalance@redhat.com>