Replace the return value from deltacloud_create_instance.
Since we are not guaranteed to get a valid document back from
the deltacloud API, we can't guarantee that we can fill in
an instance structure after a successful create_instance call.

Instead, we are guaranteed to get a Location header back from
the create_instance POST.  So replace the instance return
parameter with a char ** that holds the instance_id.  It is
then up to the library user to make an additional call to
libdeltacloud to get more information about the instance.

Signed-off-by: Chris Lalancette <clalance@redhat.com>