Cleanup libdeltacloud.syms.
After reading the ld versioning information, the way we
were doing libdeltacloud symbol versioning was wrong (I think).
What we should do is only list the version in which a symbol
was introduced, and then have later versions depend on that
version.  When we break an API (or a structure that it uses),
we move it from where it was introduced to the latest version.

Signed-off-by: Chris Lalancette <clalance@redhat.com>